backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
distro==1.9.0
dj-database-url==3.0.1
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
et_xmlfile==2.0.0
filelock==3.16.1
fsspec==2024.12.0
gunicorn==20.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.5
jiter==0.10.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.1
openai==1.95.1
openpyxl==3.1.5
pandas==2.2.3
pillow==11.1.0
psycopg2==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
sympy==1.13.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.4.0
xlrd==2.0.1
